% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gcd}
\alias{gcd}
\title{Excerpt from German Credit Data}
\format{
An object of class \code{data.frame} with 1000 rows and 4 columns.
}
\source{
\url{https://online.stat.psu.edu/stat857/node/215/}
}
\usage{
gcd
}
\description{
The German Credit Data contains data on 20 variables and the classification whether
an applicant is considered a Good or a Bad credit risk for 1000 loan applicants.
Only 3 numeric variables are extracted (Duration of Credit (month),
Credit Amount and Age (years)) along with
good/bad indicator (Creditability) and renamed as:
qual (Creditability), maturity (Duration of Credit (month)), age (Age (years)),
amount (Credit Amount).
}
\keyword{datasets}
